Added json support via spray.json Changed API provider - broker - consumer to the flexible ProjectInfo. Added support for 'global' java script files containing token and project information.
