+++ src/main/scala/org/codeprose/consumer/Consumer.scala
@@ -1,9 +1,11 @@
 package org.codeprose.consumer
 
-import org.codeprose.api.Token
+import org.codeprose.api.ProjectInfo
 
 trait Consumer {
-  def generateOutput(info:
-      scala.collection.mutable.ArrayBuffer[(java.io.File, scala.collection.mutable.ArrayBuffer[Token])]
-  ) : Unit
+  def generateOutput(projectInfo: ProjectInfo) : Unit
 }
+
+
+class ConsumerContext(
+    verbose: Boolean){}
