--- src/main/scala/org/codeprose/util/Tokenizer.scala
@@ -1,5 +1,0 @@
-package org.codeprose.util
-
-class Tokenizer {
-
-}
